refactor: move reactive stuff to its own subdir
